Move username validation helper so it can be used on the server side. Add username validation on the request_email end point
