+++ src/components/Form/Signup/Email.js
@@ -5,7 +5,7 @@ import ReCAPTCHA from 'react-google-recaptcha';
 import { Form, Icon, Input, Button } from 'antd';
 import apiCall from '../../../utils/api';
 import getFingerprint from '../../../../helpers/fingerprint';
-import { validateEmail, validateEmailDomain } from '../../../utils/validator';
+import { validateEmail, validateEmailDomain } from '../../../../helpers/validator';
 
 class Email extends React.Component {
   static contextTypes = {
